# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Talis - Celestia's infrastructure management service",
        "title": "Talis API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/instances": {
            "get": {
                "description": "Returns a list of instances with pagination and optional filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instances"
                ],
                "summary": "List instances",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of items to return (default 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to skip (default 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include deleted instances (default false)",
                        "name": "include_deleted",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by instance status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of instances",
                        "schema": {
                            "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.InstanceListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates one or more new instances",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instances"
                ],
                "summary": "Create new instances",
                "parameters": [
                    {
                        "description": "Instance creation requests",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.InstanceRequest"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created instances",
                        "schema": {
                            "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Terminates one or more instances",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instances"
                ],
                "summary": "Terminate instances",
                "parameters": [
                    {
                        "description": "Termination request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.DeleteInstancesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Instances terminated successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/instances/all-metadata": {
            "get": {
                "description": "Returns detailed metadata for all instances",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instances"
                ],
                "summary": "Get all instance metadata",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of items to return (default 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to skip (default 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include deleted instances (default false)",
                        "name": "include_deleted",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of instance metadata",
                        "schema": {
                            "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.InstanceListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/instances/public-ips": {
            "get": {
                "description": "Returns a list of public IPs for all instances",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instances"
                ],
                "summary": "Get public IPs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of items to return (default 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to skip (default 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include deleted instances (default false)",
                        "name": "include_deleted",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of public IPs",
                        "schema": {
                            "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.PublicIPsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/instances/{id}": {
            "get": {
                "description": "Returns details of a specific instance by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "instances"
                ],
                "summary": "Get instance details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Instance ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Instance details",
                        "schema": {
                            "$ref": "#/definitions/github_com_celestiaorg_talis_internal_db_models.Instance"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_celestiaorg_talis_internal_db_models.Instance": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "integer"
                },
                "payload_status": {
                    "description": "Default to PayloadStatusNone",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_celestiaorg_talis_internal_db_models.PayloadStatus"
                        }
                    ]
                },
                "project_id": {
                    "type": "integer"
                },
                "provider_id": {
                    "$ref": "#/definitions/github_com_celestiaorg_talis_internal_db_models.ProviderID"
                },
                "provider_instance_id": {
                    "type": "integer"
                },
                "public_ip": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "size": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_celestiaorg_talis_internal_db_models.InstanceStatus"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "volume_details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_celestiaorg_talis_internal_db_models.VolumeDetail"
                    }
                },
                "volume_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "github_com_celestiaorg_talis_internal_db_models.ProviderID": {
            "type": "string",
            "enum": [
                "aws",
                "gcp",
                "azure",
                "do",
                "scw",
                "vultr",
                "linode",
                "hetzner",
                "ovh",
                "do-mock",
                "digitalocean-mock",
                "mock"
            ],
            "x-enum-varnames": [
                "ProviderAWS",
                "ProviderGCP",
                "ProviderAzure",
                "ProviderDO",
                "ProviderScaleway",
                "ProviderVultr",
                "ProviderLinode",
                "ProviderHetzner",
                "ProviderOVH",
                "ProviderDOMock1",
                "ProviderDOMock2",
                "ProviderMock3"
            ]
        },
...
        "github_com_celestiaorg_talis_internal_types.DeleteInstancesRequest": {
            "type": "object",
            "required": [
                "instance_ids",
                "owner_id",
                "project_name"
            ],
            "properties": {
                "instance_ids": {
                    "description": "Instances to delete",
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "integer"
                    }
                },
                "owner_id": {
                    "description": "Owner ID",
                    "type": "integer"
                },
                "project_name": {
                    "description": "Project name",
                    "type": "string"
                }
            }
        },
        "github_com_celestiaorg_talis_internal_types.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {},
                "error": {
                    "type": "string"
                }
            }
        },
        "github_com_celestiaorg_talis_internal_types.InstanceListResponse": {
            "type": "object",
            "properties": {
                "pagination": {
                    "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.PaginationResponse"
                },
                "rows": {
                    "type": "array",
                    "items": {}
                }
            }
        },
        "github_com_celestiaorg_talis_internal_types.InstanceRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "Internal Configs - Set by the Talis Server",
                    "type": "string"
                },
                "execute_payload": {
                    "description": "Whether to execute the payload after copying",
                    "type": "boolean"
                },
                "image": {
                    "description": "OS image to use",
                    "type": "string"
                },
                "instance_id": {
                    "description": "DB Model Data - Internally set during creation",
                    "type": "integer"
                },
                "last_task_id": {
                    "description": "ID of the last task",
                    "type": "integer"
                },
                "number_of_instances": {
                    "description": "Number of instances to create",
                    "type": "integer"
                },
                "owner_id": {
                    "description": "DB Model Data - User Defined",
                    "type": "integer"
                },
                "payload_path": {
                    "description": "Local path to the payload script on the API server",
                    "type": "string"
                },
                "project_name": {
                    "description": "User Defined Configs",
                    "type": "string"
                },
                "provider": {
                    "description": "Cloud provider (e.g., \"do\")",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_celestiaorg_talis_internal_db_models.ProviderID"
                        }
                    ]
                },
                "provider_instance_id": {
                    "description": "Provider-specific instance ID",
                    "type": "integer"
                },
                "provision": {
                    "description": "Whether to run Ansible provisioning",
                    "type": "boolean"
                },
                "public_ip": {
                    "description": "Public IP address",
                    "type": "string"
                },
                "region": {
                    "description": "Region where instances will be created",
                    "type": "string"
                },
                "size": {
                    "description": "Instance size/type",
                    "type": "string"
                },
                "ssh_key_name": {
                    "description": "Name of the SSH key to use",
                    "type": "string"
                },
                "ssh_key_path": {
                    "description": "Custom path to the private SSH key file for Ansible. Overrides defaults.",
                    "type": "string"
                },
                "ssh_key_type": {
                    "description": "Talis Server Configs - Optional",
                    "type": "string"
                },
                "tags": {
                    "description": "Tags to apply to instances",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "volume_details": {
                    "description": "Detailed information about attached volumes",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.VolumeDetails"
                    }
                },
                "volume_ids": {
                    "description": "List of attached volume IDs",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "volumes": {
                    "description": "Optional volumes to attach",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.VolumeConfig"
                    }
                }
            }
        },
...
        "github_com_celestiaorg_talis_internal_types.PaginationResponse": {
            "type": "object",
            "properties": {
                "limit": {
                    "description": "Number of items per page",
                    "type": "integer"
                },
                "offset": {
                    "description": "Offset from start of results",
                    "type": "integer"
                },
                "page": {
                    "description": "Current page number",
                    "type": "integer"
                },
                "total": {
                    "description": "Total number of values",
                    "type": "integer"
                }
            }
        },
...
        "github_com_celestiaorg_talis_internal_types.PublicIPsResponse": {
            "type": "object",
            "properties": {
                "pagination": {
                    "description": "Pagination information",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.PaginationResponse"
                        }
                    ]
                },
                "public_ips": {
                    "description": "List of public IPs",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_celestiaorg_talis_internal_types.PublicIPs"
                    }
                }
            }
        },
        "github_com_celestiaorg_talis_internal_types.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
...
        "github_com_celestiaorg_talis_internal_types.VolumeDetails": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "Volume ID",
                    "type": "string"
                },
                "mount_point": {
                    "description": "Where the volume is mounted",
                    "type": "string"
                },
                "name": {
                    "description": "Volume name",
                    "type": "string"
                },
                "region": {
                    "description": "Region where volume was created",
                    "type": "string"
                },
                "size_gb": {
                    "description": "Size in gigabytes",
                    "type": "integer"
                }
            }
        },
...
    }
}